requirements.txt
Mako==1.0.0
MarkupSafe==0.23
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
decorator==3.4.0
funcsigs==0.4
matplotlib==1.3.1
neuronaut==0.0.0
nose==1.3.3
numpy==1.8.2
py==1.4.23
pycuda==2013.1.1
pyparsing==2.0.2
pytest==2.6.1
pytest-capturelog==0.7
python-dateutil==2.2
pytools==2014.3
reikna==0.6.3
six==1.7.3
tornado==4.0.1
wsgiref==0.1.2
